more TDB

+++ source/libs/tdb/src/page/tdbPage.c
@@ -19,8 +19,8 @@ extern SPageMethods pageMethods;
 extern SPageMethods pageLargeMethods;
 
 typedef struct __attribute__((__packed__)) {
-  u8 szCell[2];
-  u8 nxOffset[2];
+  u16 szCell[2];
+  u16 nxOffset[2];
 } SFreeCell;
 
 static int tdbPageAllocate(SPage *pPage, int size, SCell **ppCell);
